add columnStorage.ts
