+++ app/frontend/src/auth/authService.ts
@@ -1,0 +1,48 @@
+import { storeManager } from '../store/StoreManager';
+
+/**
+ * 認証ステータスエンドポイントを叩き、isLogin をStoreに書き込む。
+ * localhost環境ではCORSエラーになるため未ログイン扱いにする。
+ * 403もログイン済みとして扱うのは、ステージング/本番でステータスエンドポイントへの
+ * アクセス権がセッションに関係なく制限される場合があるため。
+ */
+export async function fetchLoginStatus(): Promise<void> {
+  try {
+    if (typeof window === 'undefined') {
+      storeManager.setData('isLogin', false);
+      return;
+    }
+
+    if (window.location.origin === 'http://localhost:8000') {
+      storeManager.setData('isLogin', false);
+      return;
+    }
+
+    const controller = new AbortController();
+    const timeoutId = setTimeout(() => controller.abort(), 10000);
+
+    const response = await fetch(`${window.location.origin}/auth/status`, {
+      signal: controller.signal,
+    }).catch(() => {
+      throw new Error('Request failed or timed out');
+    });
+
+    clearTimeout(timeoutId);
+
+    if (response instanceof Response) {
+      if (response.status === 200 || response.status === 403) {
+        storeManager.setData('isLogin', true);
+      } else {
+        storeManager.setData('isLogin', false);
+      }
+    }
+  } catch (error) {
+    if (
+      window.location.hostname === 'localhost' ||
+      window.location.hostname === '127.0.0.1'
+    ) {
+      console.warn('Failed to fetch auth status:', error);
+    }
+    storeManager.setData('isLogin', false);
+  }
+}
